kindergarten_management_system/requirements.txt
Django==1.8.8
django-cms==3.4.2
djangocms-text-ckeditor==3.4.0
django-filer==1.2.7
django-import-export==0.5.1
djangocms-forms==0.2.5
django-parler==1.6.4
pytz==2017.2
psycopg2
